refactor(install): extract insecure dependency policy module (review item 3)
